added snapshotlistener for db
