+++ app/src/main/java/com/example/topgmeals/MainActivity.java
@@ -1,17 +1,26 @@
 package com.example.topgmeals;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
+import android.widget.Toast;
 
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.FirebaseApp;
+import com.google.firebase.installations.FirebaseInstallations;
 
 public class MainActivity extends AppCompatActivity {
 
     // Top G
+    SharedPreferences sharedPreferences;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -19,9 +28,27 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
         FirebaseApp.initializeApp(this);
+        sharedPreferences = getSharedPreferences("Preferences", Context.MODE_PRIVATE);
+        SharedPreferences.Editor editor = sharedPreferences.edit();
 
+        FirebaseInstallations.getInstance().getId()
+                .addOnCompleteListener(new OnCompleteListener<String>() {
+                    @Override
+                    public void onComplete(@NonNull Task<String> task) {
+                        if (task.isSuccessful()) {
+                            editor.putString("Installation ID",task.getResult());
+                            editor.apply();
+                            Log.d("Installations", "Installation ID: " + task.getResult());
+                        } else {
+                            Log.e("Installations", "Unable to get Installation ID");
+                        }
+                    }
+                });
+
+        Log.d("ID", sharedPreferences.getString("Installation ID", "111"));
         Button mealPlanMealButton = (Button) findViewById(R.id.mealplan);
 
+//        Toast.makeText(this, sharedPreferences.getString("Installation ID", "111"), Toast.LENGTH_LONG).show();
         mealPlanMealButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -31,6 +58,7 @@ public class MainActivity extends AppCompatActivity {
         });
 
     }
+
 
 
 }
